decode

+++ src/decode.cpp
@@ -163,12 +163,12 @@ String getOpMode(byte input)
 
 String getInletFraction(byte input)
 {
-  return String(((input & 0b111) - 1) * 0.25);  
+  return String(((int)(input & 0b111) - 1) * 0.25);  
 }
 
 String getOutletFraction(byte input)
 {
-  return String((((input  >> 3) & 0b111) - 1) * 0.25);  
+  return String(((int)((input >> 3) & 0b111) - 1) * 0.25);  
 }
